Continue to test and implement Student aggregate changes

+++ src/Constellation.Application/Students/Events/StudentMovedSchoolsDomainEvent/UpdateOutstandingLessonRolls.cs
@@ -69,40 +69,44 @@ internal sealed class UpdateOutstandingLessonRolls
 
         List<SciencePracLesson> lessons = await _lessonRepository.GetAllForStudent(notification.StudentId, cancellationToken);
 
-        List<SciencePracRoll> oldRolls = lessons
-            .SelectMany(lesson => lesson.Rolls)
-            .Where(roll =>
-                roll.SchoolCode == notification.PreviousSchoolCode &&
-                roll.Status == LessonStatus.Active &&
-                roll.Attendance.Any(attendance => attendance.StudentId == notification.StudentId))
-            .ToList();
-
-        List<SciencePracLessonId> updatedLessonIds = new();
-
-        foreach (SciencePracRoll roll in oldRolls)
+        // The PreviousSchoolCode item may be blank if there was no active previous school when the transfer was created.
+        if (!string.IsNullOrWhiteSpace(notification.PreviousSchoolCode))
         {
-            SciencePracLesson lesson = lessons.First(lesson => lesson.Id == roll.LessonId);
+            List<SciencePracRoll> oldRolls = lessons
+                .SelectMany(lesson => lesson.Rolls)
+                .Where(roll =>
+                    roll.SchoolCode == notification.PreviousSchoolCode &&
+                    roll.Status == LessonStatus.Active &&
+                    roll.Attendance.Any(attendance => attendance.StudentId == notification.StudentId))
+                .ToList();
 
-            if (roll.Attendance.Count == 1)
+            List<SciencePracLessonId> updatedLessonIds = new();
+
+            foreach (SciencePracRoll roll in oldRolls)
             {
-                _logger.Information("Removing empty roll for lesson {lesson} at school {school}", lesson.Name, roll.SchoolCode);
+                SciencePracLesson lesson = lessons.First(lesson => lesson.Id == roll.LessonId);
 
-                // This is the last student in the class from this school
-                // The roll is no longer needed and should be cancelled
-                roll.CancelRoll("Last student has withdrawn. Roll no longer required.");
+                if (roll.Attendance.Count == 1)
+                {
+                    _logger.Information("Removing empty roll for lesson {lesson} at school {school}", lesson.Name, roll.SchoolCode);
+
+                    // This is the last student in the class from this school
+                    // The roll is no longer needed and should be cancelled
+                    roll.CancelRoll("Last student has withdrawn. Roll no longer required.");
+                }
+
+                // Remove the student from the roll
+                SciencePracAttendance? attendance = roll.RemoveStudent(notification.StudentId);
+
+                if (attendance is not null)
+                {
+                    _lessonRepository.Delete(attendance);
+                }
+
+                _logger.Information("Removing student {student} from lesson roll for {lesson} at school {school} due to moving schools", student.Name.DisplayName, lesson.Name, notification.PreviousSchoolCode);
+
+                updatedLessonIds.Add(lesson.Id);
             }
-
-            // Remove the student from the roll
-            SciencePracAttendance? attendance = roll.RemoveStudent(notification.StudentId);
-
-            if (attendance is not null)
-            {
-                _lessonRepository.Delete(attendance);
-            }
-
-            _logger.Information("Removing student {student} from lesson roll for {lesson} at school {school} due to moving schools", student.Name.DisplayName, lesson.Name, notification.PreviousSchoolCode);
-
-            updatedLessonIds.Add(lesson.Id);
         }
 
         List<SciencePracRoll> existingRollsAtNewSchool = lessons
